// lib/moviebox/generated/upstream-contract.ts
export const SIGNING = {
  secretBytes: Uint8Array.from([239,168,145,151,78,236,211,20,141,246,58,166,17,96,45,239,209,1,37,155,165,33,2,44,87,174,5,102,189,142]),
  bodyMaxBytes: 102400,
  signatureVersion: "2",
} as const;
// ...
export const FINGERPRINT = {
  "packageName": "com.community.oneroom",
  "versionName": "4.0.01.0813.03",
  "spCode": "40401",
  "userAgentTemplate": "com.community.oneroom/{} (Linux; U; Android {}; en_US; {}; Build/{}; Cronet/135.0.7012.3)",
  "versionCodes": [
    50020117,
    50020118,
    50020119,
    50020120,
    50020121
  ],
  "androidVersions": [
    [
      "9",
      "PQ3A.190605.03081104"
    ],
    [
      "10",
      "QP1A.191005.007.A3"
    ],
    [
      "11",
      "RP1A.200720.011"
    ],
    [
      "12",
      "S1B.220414.015"
    ],
    [
      "13",
      "TQ2A.230405.003"
    ]
  ],
  "devices": [
    [
      "23078RKD5C",
      "Redmi"
    ],
    [
      "2201117TY",
      "Redmi"
    ],
    [
      "2201117TG",
      "Redmi"
    ],
    [
      "22101316G",
      "Redmi"
    ],
    [
      "21121210G",
      "Redmi"
    ],
    [
      "M2012K11AG",
      "Redmi"
    ],
    [
      "M2007J20CG",
      "Redmi"
    ]
  ],
  "networkTypes": [
    "NETWORK_WIFI",
    "NETWORK_MOBILE"
  ],
  "timezones": [
    "Asia/Kolkata",
    "Asia/Shanghai",
    "Asia/Tokyo",
    "America/New_York",
    "Europe/London"
  ],
  "ipPrefixes": [
    "103.241",
    "49.36",
    "117.195",
    "106.198",
    "122.162",
    "157.32",
    "182.70",
    "103.58",
    "27.60",
    "59.90"
  ]
} as const;